requirements_databricks_ai.txt
streamlit
databricks-sql-connector
databricks-sdk
plotly
pandas==2.2.2
pyarrow==17.0.0

